chore: update README (SDKCF-3062)
