+++ Tests/DisplayPermissionServiceTests.swift
@@ -175,7 +175,7 @@ class DisplayPermissionServiceTests: QuickSpec {
 
                 it("will send required headers") {
                     preferenceRepository.setPreference(IAMPreferenceBuilder()
-                        .setAccessToken("rae-token")
+                        .setAccessToken("token")
                         .build())
 
                     waitUntil { done in
@@ -189,7 +189,7 @@ class DisplayPermissionServiceTests: QuickSpec {
                     let headers = httpSession.sentRequest?.allHTTPHeaderFields
                     expect(headers).toNot(beEmpty())
                     expect(headers?[Keys.subscriptionID]).to(equal(BundleInfoMock.inAppSubscriptionId))
-                    expect(headers?[Keys.authorization]).to(equal("OAuth2 rae-token"))
+                    expect(headers?[Keys.authorization]).to(equal("OAuth2 token"))
                 }
             }
         }
