+++ Tests/ImpressionServiceTests.swift
@@ -164,7 +164,7 @@ class ImpressionServiceTests: QuickSpec {
 
                 it("will send required headers") {
                     preferenceRepository.setPreference(IAMPreferenceBuilder()
-                        .setAccessToken("rae-token")
+                        .setAccessToken("token")
                         .build())
 
                     waitUntil { done in
@@ -179,7 +179,7 @@ class ImpressionServiceTests: QuickSpec {
                     let headers = httpSession.sentRequest?.allHTTPHeaderFields
                     expect(headers?[Keys.subscriptionID]).to(equal(BundleInfoMock.inAppSubscriptionId))
                     expect(headers?[Keys.deviceID]).toNot(beEmpty())
-                    expect(headers?[Keys.authorization]).to(equal("OAuth2 rae-token"))
+                    expect(headers?[Keys.authorization]).to(equal("OAuth2 token"))
                 }
             }
         }
